Added the basic UI for the finances

+++ client/src/services/api.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const api = axios.create({
-  baseURL: "http://localhost:5050/auth/v2",
+  baseURL: "http://localhost:5050/",
   headers: {
     "Content-Type": "application/json",
   },
@@ -40,19 +40,19 @@ api.interceptors.response.use(
 );
 const authService = {
   register: async (userData) => {
-    const response = await api.post("/signup", { ...userData });
+    const response = await api.post("/auth/v2/signup", { ...userData });
     return response;
   },
   resend: async (email) => {
-    const response = await api.post("/resend-token", { email });
+    const response = await api.post("/auth/v2/resend-token", { email });
     return response;
   },
   verify: async (token) => {
-    const response = await api.get(`/verify-email/${token}`, token);
+    const response = await api.get(`/auth/v2/verify-email/${token}`, token);
     return response;
   },
   login: async (userData) => {
-    const response = await api.post(`/login`, userData);
+    const response = await api.post(`/auth/v2/login`, userData);
     if (response.data.accessToken) {
       localStorage.setItem("access-token", response.data.accessToken);
       localStorage.setItem(
@@ -63,15 +63,17 @@ const authService = {
     return response;
   },
   forgotPass: async (email) => {
-    const response = await api.post(`/forgot-password`, { email: email });
+    const response = await api.post(`/auth/v2/forgot-password`, {
+      email: email,
+    });
     return response;
   },
   getUserDashboard: async () => {
-    const response = await api.get("/user-dashboard");
+    const response = await api.get("/auth/v2/user-dashboard");
     return response;
   },
   getAdminDashboard: async () => {
-    const response = await api.get("/get-users");
+    const response = await api.get("/auth/v2/get-users");
     return response;
   },
   getUser: () => {
@@ -80,19 +82,23 @@ const authService = {
     return user;
   },
   getCurrentUser: async () => {
-    const response = await api.get("/me");
+    const response = await api.get("/auth/v2/me");
     return response.data;
   },
   logout: async () => {
     try {
-      const response = await api.get("/logout");
+      const response = await api.get("/auth/v2/logout");
       return response;
     } finally {
       localStorage.clear();
     }
   },
   deleteUser: async (userId) => {
-    const response = await api.delete(`/delete-user/${userId}`);
+    const response = await api.delete(`/auth/v2/delete-user/${userId}`);
+    return response.data;
+  },
+  getTransactions: async () => {
+    const response = await api.get("/finances/v2/transaction");
     return response.data;
   },
 };
